Include ctor options with method options
In practice, it seems weird that calling arb.reify() will drop the
options passed to new Arborist(opts).  They can still be overridden, but
since there are no overlaps (by design), it's fine to accept them from
both places.

It may be worthwhile to move _all_ options to the ctor, and not accept
options objects in the public methods, but that's a bigger move.