Memoize many audit actions for huge speed benefits
Need to go further, though.  If you have a library with a lot of deps,
which has also published a lot of versions, and several of those deps
have advisories posted against them, then there's an exponential
explosion in the time required to figure out which versions of the
library are vulnerable.

In the case of npm's docs website, the dependency on gatsby was somewhat
punishing as a result.

We can go further in the future by caching the resulting set of
metavulns from a given set of advisory endpoint JSON, and load those
along with the initial set of vulnerabilities, so that we don't have to
scan every packument's dependencies for transitive vulnerabilities.
This commit doesn't do that, but it would be good to get that done prior
to Arborist's 1.0 release.