Avoid invalid auditReport with npm v5 lockfile
npm version 5 created package-lock.json files that did not contain the
full dependency spec in the 'requires' block.  As a result, a node may
_appear_ to be a meta-vulnerability, but when we scan for the versions
that actually _are_ vulnerable, we don't find any.

The AuditReport.delete() method was not removing the vuln from the
'effects' list of the vulnerabilities causing it, so this led to
problems when trying to print the report, and being unable to find the
missing items.

There's still an issue where audit fix won't be able to fix the problem
without first reifying to get a new lockfile, but since this only
affects projects migrating from npm v5, the impact is more minor.

cc: @ruyadorno