Refresh dep flags in virtualTree, prune extraneous from ideal
This fixes a situation where the root node's package.json is manually
edited, resulting in a package-lock.json which contains extraneous
dependencies.  If we don't mark those as extraneous, we end up reifying
them back to disk, even though they should be pruned.

This performs a fresh calcDepFlags step only in the case where the root
node's package does not match the edges defined in the package-lock.json,
to handle these cases thoroughly and completely, but only in when it's
truly necessary to do the full tree-walk.

PR-URL: https://github.com/npm/arborist/pull/119
Credit: @isaacs
Close: #119
Reviewed-by: @isaacs