Dont flag workspace edges as suspect unnecessarily
The previous update to flag edges as suspect if the package.json had
changed did not take into account the 'workspace' edges coming from the
root node.

As a result, it would _always_ flag edges as suspect if any workspaces
were in use, resulting in unnecessary work in loadVirtual.

The solution here is to load the actual workspace links when we load the
package.json for the root node, so that we can compare the actual
workspaces to the virtual workspaces present in the lockfile.

This has the side effect of requiring that we fill out the fixtures for
all the virtual workspace loading tests.  This more closely models real
world projects, however, since workspace contents are typically part of
the root project repository anyway.

This does not go as far as detecting when the package.json files of
workspace projects themselves get edited, which will likely have to be
done in a later commit.

PR-URL: https://github.com/npm/arborist/pull/122
Credit: @isaacs
Close: #122
Reviewed-by: @ruyadorno