Add strictPeerDeps, override ERESOLVE if not true
In the overwhelming majority of cases in the wild, a peer dependency
conflict that results in an ERESOLVE can be fixed by using the
`--force` flag.

However, this has other side effects (causing `npm audit fix` to install
semver-major fixes, blowing away file collisions, etc.) which might not
be desirable.  Also, since it's opt-in, it means that users have to run
the install twice for something where we're _pretty_ sure what the right
course of action is.

Let's just make that particular override the default, and reduce most
ERESOLVE errors from a crash to a warning.