fix: workspaces install entering an infinite loop
Turns out that having workspaces (or any linked packages) at the same
level depending on conflicting versions of a peer dep was not yet
handled by arborist. This specific combination was setting up the
install into an infinite loop, unable to properly resolve
build-ideal-tree._canPlaceDep.

This changeset adds tests for this specific scenario and also provides a
fix by properly resolving _canPlaceDep to a CONFLICT state if
trying to resolve conflicting versions of a peerDep within a link node,
resulting in throwing a proper ERESOLVE error in these cases.

Fixes: https://github.com/npm/cli/issues/1843