fix cycles/ordering when peer deps require nesting
Previously, there were cases where a peer dependency group might cycle
back in on itself, in such a way that we added a node and its peers
_prior_ to another member of the peer set that also peer-depended upon
the peer dependency.  Since the nodes were being moved rather than
copied, this meant that the conflicting peer was not around to be
checked, and so we would trigger peer conflict ERESOLVES when we tried
to resolve the dependencies of the second peer dependent at a higher
level.

By copying out of the virtual root peerSet, rather than moving, and
tracking which edges have already been overridden, we can avoid a lot of
other checks later on in the process, while ensuring that the
appropriate set of peer dependencies are being loaded at the optimal
level each time.