Ensure that root is added when root.meta is set
This was causing the root metadata (name, version, and the package['']
member) to be missing when we replace the metadata object on the root
node during the buildIdealTree step (which we do for 'npm update' with
no package name args).

Implicitly add the node to the metadata whenever node.meta is set.  This
required a bit of shuffling around and the removal of some cases where
we _were_ explicitly adding the root node, since it is now no longer a
thing that can be forgotten.

Fix: npm/cli#1997