Do not set fsParent to a link node
This was a weird side effect caused by the last bug, treating '..' as a
local node in the root's tree.  While that was fixed, it highlighted the
need for a better debug check on that sort of thing.

This adds a 'debug()' method, which takes a function.  The function is
ignored in normal operations, so it's perfectly fine to put expensive
and/or hazardously aggressive assertion crashes in there.  In tests,
methods passed to debug are always run, so we'll avoid creating new
scenarios where it could have been caught.