fix: missing bins info in package-lock
When migrating from package-lock v1 -> v2 in a project that currently
has an actual tree + a package-lock file the reify process fails to
properly update the package-lock with the bins info.

EDIT(@isaacs):
Added to use a somewhat questionable approach of changing the "original"
lockfile version. The questionableness comes just from the fact that the
(internal) field name says "original", which we use as an indicator as
to whether or not we can trust that this lockfile we read from disk is
actually valid or not.

In this case, after being inflated, a lockfile is as trustworthy as it
can possibly be, so there is no need to re-check and re-add package.json
files to fill in any missing pieces.

Fixes: https://github.com/npm/cli/issues/1957

PR-URL: https://github.com/npm/arborist/pull/184
Credit: @ruyadorno
Close: #184
Reviewed-by: @isaacs @nlf @darcyclarke