Refactor base classes for strict tree integrity guarantees
This is a pretty major overhaul of the entire way that nodes are added
to and removed from trees.

Essentially, the root inventory is used as an authoritative reference of
everything in the tree, and everything is updated and validated in the
setter.  So, rather than having a `parent` setter that sets the
`_parent` internal reference directly, the `parent` setter removes the
node from its root inventory, changes the path, and then calls the
`root` setter to figure out all the references.  The same with
`fsParent`, `target` and `linksIn` references between links and nodes,
and so on.

This avoids a number of hazards that have been biting us.

- Set a `x` node's parent, but don't properly connect it to nodes that have
  that parent as their `fsParent`, via
  `parent/node_modules/x/node_modules/y`.  Now, `y` will properly be
  detected as a child of `x`, and linked accordingly.
- Adding two link nodes to the same target resulting in two reachable
  nodes in the tree with the same `realpath`.  Now, it'll see that
  there's already a node with that `realpath`, and assign the link
  targets accordingly.

The list goes on and on.  This is why, for example, setting `.` as a
workspace caused some really weird data to show up in the
package-lock.json, and we had a snapshot where two `abbrev` links had a
target with a name as either `aaaaaa` or `zzzzzz`, depending on which
ended up getting added to the tree first.

The plus side is: way more consistency, with a much more forgiving API
as far as the order in which nodes get added to the tree.  The downside:
the `root` setter is a beast of delicate complexity.

Thankfully, `treeCheck(this)` is called after every call to the `root`
setter.  In debug mode, this performs an exhaustive check of the tree
and inventory, ensuring that every reachable node in the tree is present
in the inventory, that every node in the inventory is reachable, that
the entire tree has exactly one `root` node, that no nested dependencies
have `devDependencies` trying to get loaded, that no non-root nodes have
anything in their inventory objects, and so on.

This checking is of course painfully slow, and turns any error into a
fatal crash, _but_ we only run it in debug mode, which is something we
can ask users to do for debugging if they run into weird "impossible"
issues.

Caveats with this refactor:

- The root node is now included in the inventory
- The `node.location` (which is the inventory key) is the node's
  relative path based on the root's _realpath_, meaning that a Link node
  root will have a location other than `''`.

These two things together mean that we cannot test `node.location === ''`
to determine root-ness, which was done in a few places previously.  The
correct indicator is `node.isRoot`, which should always be the way to
determine whether a node is root.

This commit causes breakage in the higher level functions, which will be
fixed in subsequent commits.