Stash run scripts on arb.scriptsRun
This adds a new Set on the Arborist object called 'scriptsRun'.  It
includes objects defining which scripts have been run by the
reify/rebuild process, along with their exit status, signal, and
standard output, so that the CLI can report on it at the end of the
reify process.

PR-URL: https://github.com/npm/arborist/pull/198
Credit: @isaacs
Close: #198
Reviewed-by: @wesleytodd @ruyadorno