Preserve git+https auth when provided
This makes Arborist use the git+https: url as the saved value in
package-lock.json and package.json files for known git hosts when http
auth is provided.  When auth is not provided, we store the canonical ssh
url in package-lock.json, and the more human-friendly shortcut in
package.json, in order to maintain a canonical save value that is never
git:// for known hosts.

This is the second part of the fix for npm/cli#2054.

PR-URL: https://github.com/npm/arborist/pull/205
Credit: @isaacs
Close: #205
Reviewed-by: @wraithgar