Make printable Node/Link output a bit more useful
This tightens up the Edge representation to what is most useful,
displays Links as 'ArboristLink', and makes the code a bit easier to
work with instead of being one giant inline object.

Edges are now printed as `{ from type name@spec [error]}` for edgesIn,
and `{type name@spec [to] [error]}` for edgesOut, since we only care
about the 'from' value when it's an edge in, and only care about 'to'
when it's an edge out.  Usually, an edge will fit on one line, which is
super handy when debugging trees with lots of dependencies.
(Previously, I'd often console.error() them as
`[from.location,type,spec,to && to.location,error]` for this reason.)

It also separates out the pretty printing stuff into a standalone file, so
it's not taking up as much space in the Node class definition, and a bit
easier to maintain going forward.

PR-URL: https://github.com/npm/arborist/pull/206
Credit: @isaacs
Close: #206
Reviewed-by: @ruyadorno