More 'root is a link' cleanup
There were a few places in buildIdealTree where we do different behavior
if a node is the root.  However, it should really be if the node is
either the root, *or* the node that a Link root targets.

Add Node.isProjectRoot to mean 'the root node, or the target of a root
link node'.

PR-URL: https://github.com/npm/arborist/pull/212
Credit: @isaacs
Close: #212
Reviewed-by: @nlf