Properly track dev edges on virtual root nodes
If the virtualRoot node has a sourceReference, and that sourceReference
node would get its devDep edges loaded, then the virtualRoot node should
as well.

This caused a very useless ERESOLVE error, which could not be properly
forced away, when installing in a project that had conflicting peer deps
in its dev tree.

Related to: https://github.com/npm/cli/issues/2548