fix: package.json saving optional deps
Reify currently duplicates entries listed as optionalDependencies in the
users' package.json files. While it's working as expected this is
unexpected to a number of users and it also contradicts our own docs on
it:

    Entries in optionalDependencies will override entries of the same
    name in dependencies, so it's usually best to only put in one place.

This patches this UX problem by adding an extra check that will avoid
adding a dependency to the package.json `dependencies` object in case
that package is already listed under `optionalDependencies`.

Fixes: https://github.com/npm/cli/issues/2203
Fixes: https://github.com/npm/cli/issues/1886
Fixes: https://github.com/npm/cli/issues/724

EDIT(isaacs): Moved this into updateRootPackageJson in pairing session

PR-URL: https://github.com/npm/arborist/pull/221
Credit: @isaacs
Close: #221
Reviewed-by: @ruyadorno