Unpack shrinkwrapped deps not already unpacked
We loop over the diff.leaves to find all the shrinkwraps that _must_ be
unpacked ahead of time in order to complete the idealTree.  However, if
the idealTree already contains the children of the shrinkwrap-containing
module (because it's been previously installed and saved to a lockfile),
then we saw the hasShrinkwrap flag, and assumed it had already been
unpacked.

This tracks a Set of all nodes unpacked for the purposes of reading
their shrinkwraps, and only skips _those_ modules at unpackNewModules.

Different approach than #233, without adding extra items into the
diff.leaves, which can have the side effect of calling mkdirp more than
necessary.

Close: #233
Fix: https://github.com/npm/cli/issues/2251
Reviewed-by: @nlf