Do not rely on underscore fields in package.json
This also uncovered a bug where the yarn registry was being used when
the npm registry should have been used.  This bug is also fixed in this
commit, and was hidden by the fact that our package.json files on disk
for yarn tests were placed there by npm v6, which put the proper
resolved values in _resolved in the package.json file iteself.