do not take a link's target when removing link from tree
This causes major problems when moving links and targets between trees.
When we move the link from one tree to another, it pulls its target,
out, meaning that the target ends up becoming orphaned.

Say there is a matching link and target in two different trees, A and B,
and we want to replace the link and target in one tree with the link and
target in another.

If we move the target A to tree B first, then it replaces the target B,
and link B is updated to point to it.

Then, we move link A to tree B, replacing link B.  To do that, we move
link B out of tree B, which _takes target B with it_, leaving the ported
link A without a target in tree B.