fix(add): save packages in the right place
Right now where we end up saving a dep is not very well coupled to where
the user requested it.  We are inferring more than we are checking for
intention.  This moves us to more heavily lean on the `saveType` if
given when adding a package into the dependencies.

We also were not consistently inferring the saveType.  There was code in
two places trying to do this, and either place used slightly different
logic. This moves the logic into one place.

Finally, we are more consistently making certain dependency types
mutually exclusive, for instance dependencies and devDependencies.

PR-URL: https://github.com/npm/arborist/pull/273
Credit: @wraithgar
Close: #273
Reviewed-by: @nlf