fix(reify): do not clean up nodes with no parent
Co-authored by @nlf

When reify goes to add omits to the trash list, sometimes it adds things
it shouldn't, such as top level nodes that are actually links, but
they're only links elsewhere in the tree where they are properly placed
(i.e. with the root as the parent).

This PR prevents things like "npm install --production" deleting the
contents of a linked dev dependency

PR-URL: https://github.com/npm/arborist/pull/277
Credit: @wraithgar
Close: #277
Reviewed-by: @nlf