fix(reify): properly save spec if prerelease
Currently if you add a package w/ spec `*` (or with no spec, which
defaults to `*`) arborist tries to do a semver.subset to see if it can
use the spec that actually got installed. This fails if the new spec has
a prerelease tag in it.

This change fixes it so that if the spec was `*` it just goes ahead and
uses the new spec no matter what, since no valid semver doesn't match
`*` in this case